[BREAKING] Recreated all migrations
The number of migrations was getting too significant to easily maintain,
so I've re-created all tables in a manner that works with MySQL and
SQLite and which re-structures some fields that could not be moved
before.

Obviously, this is a severly breaking change. Databases will need to be
backed up beforehand.